converts all Int types to Long to support large transactions in satoshis

+++ src/main/scala/com/github/scala_opennode/entities/ChainInvoice.scala
@@ -11,6 +11,6 @@ object ChainInvoice {
  */
 case class ChainInvoice(
     address: String, // charge charges
-    settled_at: Option[Int], // charges
+    settled_at: Option[Long], // charges
     tx: Option[String] // charges
 )
